requirements.txt
pandas==1.5.3
numpy==1.23.5
matplotlib==3.6.2
seaborn==0.12.2
scikit-learn==1.2.1
pyyaml==6.0
loguru==0.7.0
